DynamicObject float -> double

+++ src/DynamicObject.hpp
@@ -22,8 +22,8 @@
 #include <functional>
 #include <vector>
 
-typedef std::function<float(const std::vector<glm::vec3> &)> constraint_function;
-typedef std::function<std::vector<glm::vec3>(const std::vector<glm::vec3> &)> gradient_function;
+typedef std::function<double(const std::vector<glm::dvec3> &)> constraint_function;
+typedef std::function<std::vector<glm::dvec3>(const std::vector<glm::dvec3> &)> gradient_function;
 
 enum ConstraintType {
     EQUALITY_CONSTRAINT,
@@ -32,12 +32,12 @@ enum ConstraintType {
 
 class DynamicObject {
     // Verticies
-    uint N = 0;                          // number of vertices
-    std::vector<glm::vec3> m_positions;  // xi
-    std::vector<glm::vec3> m_velocities; // vi
-    std::vector<float> m_masses;         // mi
-    std::vector<float> m_weights;        // wi
-    std::vector<bool> m_fixed;           // if the vertex is fixed
+    uint N = 0;                           // number of vertices
+    std::vector<glm::dvec3> m_positions;  // xi
+    std::vector<glm::dvec3> m_velocities; // vi
+    std::vector<double> m_masses;         // mi
+    std::vector<double> m_weights;        // wi
+    std::vector<bool> m_fixed;            // if the vertex is fixed
 
     // Constraints
     uint M = 0;                                   // number of contraints
@@ -45,11 +45,11 @@ class DynamicObject {
     std::vector<constraint_function> m_functions; // Cj: The constraint itself. Input's size must match the cardinality
     std::vector<gradient_function> m_gradients;   // Cj: The gradient (evolution) of the constraint. Input's size must match the cardinality
     std::vector<std::vector<uint>> m_indices;     // Indices of impacted vertices
-    std::vector<float> m_stiffnesses;             // kj: Strength in [0;1]
+    std::vector<double> m_stiffnesses;            // kj: Strength in [0;1]
     std::vector<ConstraintType> m_types;          // Either Equality (=0) or Inequality (>=0)
 
     // "3.5. Damping" of ./articles/Position_Based_Dynamics.pdf
-    void dampVelocities(float k_damping = 1.f); // k_damping = 1. -> rigid body
+    void dampVelocities(double k_damping = 1.); // k_damping = 1. -> rigid body
 
     void fillMissingVertexInfos() {
         m_velocities.resize(N);
@@ -59,27 +59,27 @@ class DynamicObject {
 public:
     // GETTERS
     uint getN() const { return N; };
-    const std::vector<glm::vec3> &getPositions() const { return m_positions; };
-    const std::vector<glm::vec3> &getVelocities() const { return m_velocities; };
-    const std::vector<float> &getMasses() const { return m_masses; };
-    const std::vector<float> &getWeights() const { return m_weights; };
+    const std::vector<glm::dvec3> &getPositions() const { return m_positions; };
+    const std::vector<glm::dvec3> &getVelocities() const { return m_velocities; };
+    const std::vector<double> &getMasses() const { return m_masses; };
+    const std::vector<double> &getWeights() const { return m_weights; };
     const std::vector<bool> &getFixed() const { return m_fixed; };
     uint getM() const { return M; };
     const std::vector<uint> &getCardinalities() const { return m_cardinalities; };
     const std::vector<constraint_function> &getFunctions() const { return m_functions; };
     const std::vector<gradient_function> &getGradients() const { return m_gradients; };
     const std::vector<std::vector<uint>> &getIndices() const { return m_indices; };
-    const std::vector<float> &getStiffnesses() const { return m_stiffnesses; };
+    const std::vector<double> &getStiffnesses() const { return m_stiffnesses; };
     const std::vector<ConstraintType> &getTypes() const { return m_types; };
 
     // "3.1. Algorithm Overview" of ./articles/Position_Based_Dynamics.pdf
-    void update(float _delta_time);
+    void update(double _delta_time);
 
-    void addVertex(const glm::vec3 &_position, const glm::vec3 &_velocity, float _mass, bool _fixed);
-    void setVertexPosition(uint _pj, glm::vec3 _position) { m_positions[_pj] = _position; }
-    void setVertexVelocity(uint _pj, glm::vec3 _velocity) { m_velocities[_pj] = _velocity; }
-    void setVertexMass(uint _pj, float _mass) { m_masses[_pj] = _mass; }
-    void setVertexWeight(uint _pj, float _weight) { m_weights[_pj] = _weight; }
+    void addVertex(const glm::dvec3 &_position, const glm::dvec3 &_velocity, double _mass, bool _fixed);
+    void setVertexPosition(uint _pj, glm::dvec3 _position) { m_positions[_pj] = _position; }
+    void setVertexVelocity(uint _pj, glm::dvec3 _velocity) { m_velocities[_pj] = _velocity; }
+    void setVertexMass(uint _pj, double _mass) { m_masses[_pj] = _mass; }
+    void setVertexWeight(uint _pj, double _weight) { m_weights[_pj] = _weight; }
     void setVertexFixed(uint _pj, bool _fixed);
 
     void addConstraint(
@@ -87,12 +87,12 @@ public:
         const constraint_function &_function,
         const gradient_function &_gradient,
         const std::vector<uint> &_indices,
-        float _stiffness,
+        double _stiffness,
         const ConstraintType &_type);
-    void addDistanceConstraint(uint _p0, uint _p1, float _stiffness, float _targeted_distance);
-    void addDistanceConstraint(uint _p0, uint _p1, float _stiffness); // the targeted distance is set to the current distance between p0 and p1
-    void addBendingConstraint(uint _p0, uint _p1, uint _p2, uint _p3, float _stiffness, float _targeted_angle);
-    void addBendingConstraint(uint _p0, uint _p1, uint _p2, uint _p3, float _stiffness); // the targeted angle is set to the current angle between p0,p2,p1 normal and p0,p3,p1 normal
+    void addDistanceConstraint(uint _p0, uint _p1, double _stiffness, double _targeted_distance);
+    void addDistanceConstraint(uint _p0, uint _p1, double _stiffness); // the targeted distance is set to the current distance between p0 and p1
+    void addBendingConstraint(uint _p0, uint _p1, uint _p2, uint _p3, double _stiffness, double _targeted_angle);
+    void addBendingConstraint(uint _p0, uint _p1, uint _p2, uint _p3, double _stiffness); // the targeted angle is set to the current angle between p0,p2,p1 normal and p0,p3,p1 normal
 
     // OpenGL interface
